changes per PR feedback
